# swagger.yaml
openapi: 3.0.1
info:
  title: Vertragsmanagement
  version: '1.0'
paths:
...
  /bankverbindungen/search:
    get:
      tags:
        - Bankverbindungen
      summary: Suche nach Bankverbindungen anhand von Schlagwörtern
      parameters:
        - name: searchString
          in: query
          description: Sucheingabe
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bankverbindung'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bankverbindung'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bankverbindung'
        '418':
          description: I'm a teapot. Empty search string.
  /bankverbindungen/{iban}:
    get:
      tags:
        - Bankverbindungen
      summary: Gibt die Zahlungsinformationen anhand einer IBAN aus
      parameters:
        - name: iban
          in: path
          description: IBAN der Bankverbindung
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success. Return IBAN.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Bankverbindung'
            application/json:
              schema:
                $ref: '#/components/schemas/Bankverbindung'
            text/json:
              schema:
                $ref: '#/components/schemas/Bankverbindung'
        '404':
          description: Not found
    delete:
      tags:
        - Bankverbindungen
      summary: Löscht die Bankverbindung
      parameters:
        - name: iban
          in: path
          description: IBAN der Bankverbindung
          required: true
          schema:
            type: string
        - name: force
          in: query
          description: Force-Parameter
          schema:
            type: boolean
      responses:
        '200':
          description: Success
...
  /debitoren/search:
    get:
      tags:
        - Debitoren
      summary: Suche nach Debitoren anhand von Schlagwörtern
      parameters:
        - name: searchString
          in: query
          description: Sucheingabe
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Debitor'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Debitor'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Debitor'
        '418':
          description: I'm a teapot. Empty search string.
  /debitoren/{debitorId}:
    get:
      tags:
        - Debitoren
      summary: Gibt Debitoren anhand von der Debitoren-Nummer aus
      parameters:
        - name: debitorId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success. Return Debitor.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Debitor'
            application/json:
              schema:
                $ref: '#/components/schemas/Debitor'
            text/json:
              schema:
                $ref: '#/components/schemas/Debitor'
        '404':
          description: Not found
  /debitoren/{debitorId}/adresse:
    get:
      tags:
        - Debitoren
      summary: Gibt Adresse anhand einer Debitoren-Nummer aus
      parameters:
        - name: debitorId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success. Return Adresse.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Adresse'
            application/json:
              schema:
                $ref: '#/components/schemas/Adresse'
            text/json:
              schema:
                $ref: '#/components/schemas/Adresse'
        '404':
          description: Not found
        '500':
          description: Database error
...
  /kreditoren/search:
    get:
      tags:
        - Kreditoren
      summary: Suche nach Kreditoren anhand von Schlagwörtern
      parameters:
        - name: searchString
          in: query
          description: Sucheingabe
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Kreditor'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Kreditor'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Kreditor'
        '418':
          description: I'm a teapot. Empty search string.
  /kreditoren/{kreditorId}:
    get:
      tags:
        - Kreditoren
      summary: Gibt Kreditoren anhand von der Kreditoren-Nummer aus
      parameters:
        - name: kreditorId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success. Return Kreditor.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Kreditor'
            application/json:
              schema:
                $ref: '#/components/schemas/Kreditor'
            text/json:
              schema:
                $ref: '#/components/schemas/Kreditor'
        '404':
          description: Not found
  /kreditoren/{kreditorId}/adresse:
    get:
      tags:
        - Kreditoren
      summary: Gibt Adresse anhand einer Kreditoren-Nummer aus
      parameters:
        - name: kreditorId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success. Return Adresse.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Adresse'
            application/json:
              schema:
                $ref: '#/components/schemas/Adresse'
            text/json:
              schema:
                $ref: '#/components/schemas/Adresse'
        '404':
          description: Not found
        '500':
          description: Database error
...
components:
  schemas:
    Adresse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        adresszeile1:
          type: string
          nullable: true
        adresszeile2:
          type: string
          nullable: true
        hausnummer:
          type: string
          nullable: true
        postleitzahl:
          type: string
          nullable: true
        ort:
          type: string
          nullable: true
        inUse:
          type: integer
          format: int32
      additionalProperties: false
    Bankverbindung:
      required:
        - kontoinhaber
      type: object
      properties:
        iban:
          type: string
          nullable: true
        bic:
          type: string
          nullable: true
        kontoinhaber:
          minLength: 1
          type: string
        inUse:
          type: integer
          format: int32
      additionalProperties: false
    Debitor:
      required:
        - bankverbindung
        - isCustomer
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          minLength: 1
          type: string
        vorname:
          type: string
          nullable: true
        adresse:
          type: integer
          format: int32
        bankverbindung:
          minLength: 1
          type: string
        isCustomer:
          type: boolean
      additionalProperties: false
    Kreditor:
      required:
        - isCustomer
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          minLength: 1
          type: string
        vorname:
          type: string
          nullable: true
        adresse:
          type: integer
          format: int32
        bankverbindung:
          type: string
          nullable: true
        isCustomer:
          type: boolean
      additionalProperties: false
...